backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pandas==2.1.3
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
passlib==1.7.4
python-jose==3.3.0
email-validator==2.0.0
bcrypt==4.0.1
